Fixed duplication issue with catwalk_anovas
Running the same anova analysis multiple times using catwalk_anovas produced multiple identical copies of the analysis. The function was changed to check if an analysis using the same data, formula, parameters, etc has already been completed, and throw a warning asking the user if they want to produce duplicated output or not and how to do so if they do.